[DS-514] logic for index creation identifier spec

+++ schematools/importer/base.py
@@ -4,7 +4,7 @@ from itertools import islice
 import operator
 from typing import Optional, Dict
 from jsonpath_rw import parse
-from sqlalchemy import MetaData, inspect, Table, Column, ForeignKey, Integer, String
+from sqlalchemy import MetaData, inspect, Table, Column, ForeignKey, Integer, String, Index, exc #adjusted
 
 from schematools import MAX_TABLE_LENGTH
 from schematools.types import DatasetSchema, DatasetTableSchema
@@ -147,37 +147,46 @@ class BaseImporter:
     def create_pk_lookup(self, tables):
         """ Generate a lookup to avoid primary_key clashes """
         for table_name, table in tables.items():
-            pk_columns = inspect(table).primary_key.columns
-            # nm tables do not have a PK
-            if not pk_columns:
-                return
-            # We assume a single PK (because of Django)
-            pk_col = pk_columns.values()[0]
-            pk_name = pk_col.name
-            if pk_col.autoincrement == "auto":
-                continue
-            self.pk_colname_lookup[table_name] = pk_name
-            pks = set(
-                [getattr(r, pk_name) for r in self.engine.execute(table.select())]
-            )
-            self.pk_values_lookup[table_name] = pks
+            if isinstance(table, Table): #adjusted
+                pk_columns = inspect(table).primary_key.columns
+                # nm tables do not have a PK
+                if not pk_columns:
+                    return
+                # We assume a single PK (because of Django)
+                pk_col = pk_columns.values()[0]
+                pk_name = pk_col.name
+                if pk_col.autoincrement == "auto":
+                    continue
+                self.pk_colname_lookup[table_name] = pk_name
+                pks = set(
+                    [getattr(r, pk_name) for r in self.engine.execute(table.select())]
+                )
+                self.pk_values_lookup[table_name] = pks
 
     def generate_tables(self, table_name, db_table_name=None, truncate=False):
         """Generate the tablemodels and tables"""
 
-        self.dataset_table = self.dataset_schema.get_table_by_id(table_name)
+        self.dataset_table = self.dataset_schema.get_table_by_id(table_name)    
+
         # Collect provenance info for easy re-use
         self.fields_provenances = self.fetch_fields_provenances(self.dataset_table)
         self.db_table_name = db_table_name
         if db_table_name is None:
             self.db_table_name = get_table_name(self.dataset_table)
         # Bind the metadata
-        metadata.bind = self.engine
+        metadata.bind = self.engine        
         # Get a table to import into
         self.tables = table_factory(
             self.dataset_table, metadata=metadata, db_table_name=self.db_table_name
         )
+        
+        # inspect on metadata.bind needed for index creation (prepare_identifier_index)
+        try:#adjusted
+            metadata_inspector = inspect(metadata.bind)       
+        except exc.NoInspectionAvailable as e:
+            metadata_inspector = []
         self.prepare_tables(self.tables, truncate=truncate)
+        self.prepare_identifier_index(self.tables, metadata_inspector, metadata.bind)#adjusted
         self.create_pk_lookup(self.tables)
 
     def load_file(
@@ -231,13 +240,53 @@ class BaseImporter:
         raise NotImplementedError()
 
     def prepare_tables(self, tables, truncate=False):
-        """Create the tables if needed"""
-        for table in tables.values():
-            if not table.exists():
-                table.create()
-            elif truncate:
-                self.engine.execute(table.delete())
+        """Create the tables if needed""" 
 
+        for table in tables.values():
+            if isinstance(table, Table): #adjusted
+                                           
+                if not table.exists():
+                    table.create()
+                    
+                elif truncate:
+                    self.engine.execute(table.delete())
+        
+    def prepare_identifier_index(self, tables, inspector, engine):
+        """ Create indexes if needed """
+
+        table_name = None
+        table_object = None
+        ind_identifier_index = False
+
+        for object_name, object in tables.items():#adjusted  
+
+            if isinstance(object, Table): 
+                # get the table name and Table instance
+                # the Table instance is assigned to the generic var name "table_object"
+                # which is referenced by the Index SQLalchemy class instantiation and 
+                # expects a Table instance as it's value  
+                table_name = object_name
+                table_object = object
+            
+            # check for presence of the _identifier_idx Index object
+            if object_name == f'{table_name}_identifier_idx':
+       
+                try: 
+                    # fetch indexes based on specification in DB
+                    indexes = inspector.get_indexes(table_name, schema=None)                    
+                    for index in indexes:
+                        if index.get('name', None) == object_name:
+                            ind_identifier_index = True
+
+                    if not ind_identifier_index:
+                        print("Identifier index not found...start creating", object)
+                        # eval is used to execute string to code (which is used to make it dynamicly generated)
+                        eval(object).create(bind=engine)                            
+
+                except AttributeError as e:
+                    print(f"Error on identifier index for {object_name}:", e)
+                    continue            
+          
 
 class CliLogger:
     def __index__(self, batch_size):
@@ -299,6 +348,8 @@ def table_factory(
     metadata = metadata or MetaData()
     sub_tables = {}
     columns = []
+    identifier_index = {}
+    
     for field in dataset_table.fields:
         if field.type.endswith("#/definitions/schema"):
             continue
@@ -314,7 +365,8 @@ def table_factory(
                     sub_columns = [
                         Column("id", Integer, primary_key=True),
                         Column("parent_id", ForeignKey(fk_column, ondelete="CASCADE")),
-                    ]
+                    ]             
+                
 
                 elif field.is_through_table:
                     # We need a 'through' table for the n-m relation
@@ -359,6 +411,11 @@ def table_factory(
                 continue
 
             col_type = fetch_col_type(field)
+
+            if dataset_table.identifier and field.name in dataset_table.identifier: #adjusted                
+                identifier_index = define_identifier_index(db_table_name, dataset_table)
+
+            
         except KeyError:
             raise NotImplementedError(
                 f'Import failed at "{field.name}": {dict(field)!r}\n'
@@ -370,8 +427,42 @@ def table_factory(
             col_kwargs["primary_key"] = True
             col_kwargs["nullable"] = False
             col_kwargs["autoincrement"] = False
-
+        
+        
         id_postfix = "_id" if field.relation else ""
-        columns.append(Column(f"{field_name}{id_postfix}", col_type, **col_kwargs))
+        columns.append(Column(f"{field_name}{id_postfix}", col_type, **col_kwargs))    
+    
+    
+    return {db_table_name: Table(db_table_name, metadata, *columns), **identifier_index, **sub_tables} 
 
-    return {db_table_name: Table(db_table_name, metadata, *columns), **sub_tables}
+    
+
+def define_identifier_index(db_table_name: str, dataset_table: DatasetTableSchema) -> Dict: #adjusted   
+    """Generate one or more SQLAlchemy Index objects to work with the JSON Schema
+
+    :param db_table_name: The full table name: [name_dataset]_ [name_table]
+    :param dataset_table: The Amsterdam Schema definition of the table
+    
+    In the JSON Schema definition of the table, an identifier arry maybe definied.
+    I.e. "identifier": ["identificatie", "volgnummer"]
+    
+    The purpose is to set, besides the (single) PK, the (compound) Alternate Key to uniquely identify a table record.
+    This is frequently used in temporal tables where versions of data are present (history).
+
+    An index on these AK is needed in order to maximize performance when SQL joining to the vesionized record.
+    """
+    identifier_column_snaked = []
+    index = {} 
+
+    if not identifier_column_snaked:   
+                
+        # precautionary measure: If camelCase, translate to snake_case, so column(s) can be found in table
+        # table_object is a generic name which contains a instance of SQLalchemy Table class, which references to the DB table
+        # the letter 'c' between table_object and field of identifier, is a mandatory element of SQLalchemy to specifiy the 
+        # fields that will be part of the index creation.
+        for identifier_column in dataset_table.identifier:                   
+            identifier_column_snaked.append(f'table_object.c.{to_snake_case(identifier_column)}')             
+
+        index[f'{db_table_name}_identifier_idx'] = f"Index('{db_table_name}_identifier_idx', {','.join(identifier_column_snaked)})"
+    
+    return index
